some interface adjustemns for the cash desk and impl of adapters

+++ Domain/CashDesk/IBankService.cs
@@ -2,7 +2,7 @@
 
 public interface IBankService
 {
-    Task<BankTransactionContext> CreateTransactionContextAsync(double amount);
+    Task<BankTransactionContext> CreateTransactionContextAsync(int amount);
     
     Task<AuthorizationResult> AuthorizePaymentAsync(BankTransactionContext context, string token);
